update concept handling based on proper authorization🚀

+++ app/controllers/manage/concepts_controller.ts
@@ -1,9 +1,14 @@
 import type { HttpContext } from '@adonisjs/core/http'
 import Concept from '#models/concept'
 import { createConceptValidator, updateConceptValidator } from '#validators/concept'
+import ConceptPolicy from '#policies/concept_policy'
 
-export default class ConceptsController {
-  async index({ response }: HttpContext) {
+export default class ManageConceptsController {
+  async index({ bouncer, response }: HttpContext) {
+    if (await bouncer.with(ConceptPolicy).denies('view')) {
+      return response.forbidden('Cannot view this page')
+    }
+
     const concepts = await Concept.query()
       .preload('children')
       .whereNull('parent_id')
@@ -12,10 +17,14 @@ export default class ConceptsController {
     return response.json(concepts)
   }
 
-  async store({ request, response }: HttpContext) {
+  async store({ bouncer, request, response }: HttpContext) {
     try {
       const data = await request.validateUsing(createConceptValidator)
       const concept = await Concept.create(data)
+
+      if (await bouncer.with(ConceptPolicy).denies('create', concept)) {
+        return response.forbidden('Cannot create this concept')
+      }
 
       await concept.load('children')
       return response.created(concept)
@@ -24,7 +33,11 @@ export default class ConceptsController {
     }
   }
 
-  async show({ params, response }: HttpContext) {
+  async show({ bouncer, params, response }: HttpContext) {
+    if (await bouncer.with(ConceptPolicy).denies('view')) {
+      return response.forbidden('Cannot view this page')
+    }
+
     const concept = await Concept.findByOrFail('slug', params.slug)
     await concept.load('children')
 
@@ -39,10 +52,14 @@ export default class ConceptsController {
     return response.json(concept)
   }
 
-  async update({ params, request, response }: HttpContext) {
+  async update({ bouncer, params, request, response }: HttpContext) {
     try {
       const concept = await Concept.findByOrFail('slug', params.slug)
       const data = await request.validateUsing(updateConceptValidator)
+
+      if (await bouncer.with(ConceptPolicy).denies('update', concept)) {
+        return response.forbidden('Cannot update this concept')
+      }
 
       await concept.merge(data).save()
       await concept.load('children')
@@ -52,9 +69,15 @@ export default class ConceptsController {
     }
   }
 
-  //   async updateContent({ params, request, response }: HttpContext) {
+  // This controller is specific for a knowledge block
+
+  //   async updateContent({ bouncer, params, request, response }: HttpContext) {
   //     try {
   //       const concept = await Concept.findByOrFail('slug', params.slug)
+
+  //     if (await bouncer.with(ConceptPolicy).denies('update', concept)) {
+  //       return response.forbidden('Cannot update this concept')
+  //     }
 
   //       if (!concept.isTerminal) {
   //         return response.badRequest({
@@ -71,8 +94,12 @@ export default class ConceptsController {
   //     }
   //   }
 
-  async destroy({ params, response }: HttpContext) {
+  async destroy({ bouncer, params, response }: HttpContext) {
     const concept = await Concept.findByOrFail('slug', params.slug)
+
+    if (await bouncer.with(ConceptPolicy).denies('update', concept)) {
+      return response.forbidden('Cannot update this concept')
+    }
     await concept.delete()
 
     return response.noContent()
